Split aspects in different modules
Now it's easier to reason about each individual aspect
