+++ gaphas/painter/focuseditempainter.py
@@ -1,7 +1,27 @@
-from typing import Collection
+from __future__ import annotations
 
-from gaphas.aspect import PaintFocused
+from functools import singledispatch
+from typing import TYPE_CHECKING, Collection
+
 from gaphas.item import Item
+
+if TYPE_CHECKING:
+    from gaphas.view import GtkView
+
+
+class ItemPaintFocused:
+    """Paints on top of all items, just for the focused item and only when it's
+    hovered (see gaphas.painter.FocusedItemPainter)"""
+
+    def __init__(self, item: Item, view: GtkView):
+        self.item = item
+        self.view = view
+
+    def paint(self, cairo):
+        pass
+
+
+PaintFocused = singledispatch(ItemPaintFocused)
 
 
 class FocusedItemPainter:
