+++ gaphas/types.py
@@ -1,0 +1,3 @@
+from typing import Tuple
+
+Pos = Tuple[float, float]
